docs(larastan): el nivel 6 no encuentra nada, y el 7 sí
La escalera de phpstan se llevaba subiendo por inercia y funcionó cinco veces
seguidas. La sexta no. Medido antes de tocar nada: el nivel 6 son 1.940
errores, el 100% anotaciones de tipo que faltan, y **ninguno señala código que
pueda fallar** — se revisaron a mano los 24 que no eran anotación pura por si
escondían la forma del count() sobre un Builder de la §13.1, y son
collect($r->json()) en los tests. Además 1.322 de los 1.940 caen en los 129
controladores: el mismo diff ilegible que CLAUDE.md prohíbe para Pint, y por el
mismo motivo.

Donde está el valor es en el 7, y no por ser más alto sino porque es otra
familia de reglas: method.nonObject, property.nonObject, foreach.nonIterable,
offsetAccess.nonOffsetAccessible, assign.propertyType y return.type son
«se usa como objeto algo que no lo es», que es exactamente lo que fueron los
dos hallazgos más caros de la serie —el count() sobre Builder y el
$datos->id sobre un array—. Son 24 en todo el repo.

De ahí salen los arreglos de este commit:

- TokenDeSesion::segundosDeVida() emitía un «Implicit conversion from float to
  int loses precision» de PHP 8.1 en CADA login y cada refresco, porque Carbon 3
  devuelve float y expires_at casi nunca cae en un segundo entero. Código
  nuestro, de la migración, que ni los 438 tests ni el nivel 5 veían: un
  deprecated no rompe nada, solo llena el log de los dieciséis colegios.
- Services\Login::entrar() no era un fallo —$fila->id sale de un SELECT u.id—
  pero el (int) explícito cierra la rama en que find() devolvería una Collection.
- Tres foreach sobre lo que puede ser false en tools/, de una línea cada uno.

El bloque para subir el gate a 7 queda escrito y SIN poner: cambia
`composer run stan` para las tres sesiones que trabajan hoy sobre este árbol, y
18 de los 24 errores están en ficheros que las otras dos tienen abiertos. Van
traspasados y medidos en la §5 del documento.

+++ app/Models/TokenDeSesion.php
@@ -67,6 +67,10 @@ class TokenDeSesion extends PersonalAccessToken
             return 0;
         }
 
-        return max(0, Carbon::now()->diffInSeconds($this->expires_at, false));
+        // `(int)` explícito: `diffInSeconds` devuelve **float** en Carbon 3, y
+        // `expires_at` casi nunca cae en un segundo entero, así que la conversión
+        // implícita del `: int` emitía un «Implicit conversion from float ... loses
+        // precision» de PHP 8.1 en **cada** login y cada refresco.
+        return (int) max(0, Carbon::now()->diffInSeconds($this->expires_at, false));
     }
 }
